Run jobs on standbys
One difficulty in running jobs in standbys is that you can't have
logging of run details to tables.

But their executions can still be recorded in the database audit logs,
and they can also be perceived by external locations, when they have
side effects through network requests or FDW and such.

So it's justifyable to support use cases where logging to tables isn't a
vital necessity.

This adds such a possibility by making the standby read the cron.job
table in search of jobs that match its state, and not recording their
runs in the details table.

Since there are certainly jobs that should be only executed on the
primary, as well as jobs that should be ran only on the standby, and
jobs that may be ran in both types of instance, this adds a new
attribute 'scope' to cron.job, with possible values 'primary',
'standby', and 'both', where 'primary' is the default, to match the
established behavior.

This also bumps the extension version to 1.7 and provides new function
signatures to schedule and alter jobs. Due to the various arguments with
default expressions in the v1.6 functions, we could either drop them and
recreate, losing the GRANTs that users might have on them, OR create
entirely new function names to include scope as argument, OR overload
the functions. Overloading was the choice this time, but it brings the
issue of ambiguous function resolutions, so we could only overload using
functions that have no default expressions. We may revisit this
decision.

This also adds a TAP test to cover the new functionality.